Web build pipeline cleanup + browser UX fixes + web documentation
Summary of changes

Removed web-release preset and kept web-release-split as the release web mode.
Updated build_web_html.bat:
default preset is now web-release-split
release is now an alias to release-split
usage text updated accordingly
Updated web shell (app_resources/web/index_shell.html) to suppress browser context menu globally so RMB attack works without popup.
Fixed clean.bat to match current repo layout:
removed invalid pushd game logic
cleans root build folders and CMake artifacts reliably
Updated docs:
sofaspuds_ReadMe.txt and projectGuide.txt now include web prerequisites and run steps
added webFeatureGuide.txt with detailed architecture/build/runtime/troubleshooting notes
Testing / verification

build_web_html.bat release-split reconfigure completes.
python -m http.server 8000 -d build\web-release-split\Sandbox serves correctly.
BloodyGoodCurry.html loads with .js/.wasm/.data.
Right-click no longer opens browser menu in gameplay.
Docs reflect current web workflow and required tooling.

+++ Sandbox/MyGame/Scripts/GameScripts.cpp
@@ -587,8 +587,8 @@ namespace {
             aimDirX = state.lastAimDirX;
             aimDirY = state.lastAimDirY;
         }
-        state.lastMouseX = mouse.x;
-        state.lastMouseY = mouse.y;
+        state.lastMouseX = static_cast<float>(mouse.x);
+        state.lastMouseY = static_cast<float>(mouse.y);
         
         /*************************************************************************************
           \brief Flips the sprite based on final aim direction, for both mouse and controller
